+++ src/sentry/android/android_sdk.h
@@ -21,6 +21,21 @@ protected:
 	static void _bind_methods();
 };
 
+class SentryAndroidBeforeSendFeedbackHandler : public Object {
+	GDCLASS(SentryAndroidBeforeSendFeedbackHandler, Object);
+	friend class AndroidSDK;
+
+private:
+	Object *android_plugin = nullptr;
+
+	void _initialize(Object *p_android_plugin) { android_plugin = p_android_plugin; }
+
+	void _before_send_feedback(int32_t p_event_handle);
+
+protected:
+	static void _bind_methods();
+};
+
 class SentryAndroidBeforeSendLogHandler : public Object {
 	GDCLASS(SentryAndroidBeforeSendLogHandler, Object);
 	friend class AndroidSDK;
@@ -56,6 +71,7 @@ class AndroidSDK : public InternalSDK {
 private:
 	uint64_t android_plugin_instance_id = 0;
 	SentryAndroidBeforeSendHandler *before_send_handler = nullptr;
+	SentryAndroidBeforeSendFeedbackHandler *before_send_feedback_handler = nullptr;
 	SentryAndroidBeforeSendLogHandler *before_send_log_handler = nullptr;
 	SentryAndroidBeforeSendMetricHandler *before_send_metric_handler = nullptr;
 
@@ -98,6 +114,7 @@ public:
 	virtual SentryScopeImpl *create_scope() override;
 
 	virtual bool supports_scopes() const override { return true; }
+	virtual bool supports_before_send_feedback() const override { return true; }
 
 	virtual void set_trace(const String &p_trace_id, const String &p_parent_span_id) override;